+++ apps/factory-api/src/modules/menus/menus.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common'
 
+import { ProfilesModule } from '../profiles/profiles.module'
 import { MenusController } from './menus.controller'
 import { MenusService } from './menus.service'
 
 @Module({
+  imports: [ProfilesModule],
   controllers: [MenusController],
   providers: [MenusService],
 })