+++ apps/factory-api/src/modules/profiles/profiles.module.ts
@@ -6,5 +6,6 @@ import { ProfilesService } from './profiles.service'
 @Module({
   controllers: [ProfilesController],
   providers: [ProfilesService],
+  exports: [ProfilesService],
 })
 export class ProfilesModule {}